+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## UNRELEASED
+
+### Adds
+
+* Supports body style in and out of media queries. Also supports body style applied to indentifiers like ids and classes if they exist on the body tag. 
+Uses the new `[data-apos-refreshable-body]` that exists and replace the body in breakpoint preview in core.
+
 ## 1.1.0 (2025-03-19)
 
 ### Adds